Add exponential moving average and isApogee
Adds a function for exponential moving average and isApogee.
Exponential moving average puts in the unfiltered altitude with fewer
than numValues values, which is currently set to 20. It puts in the SMA
for the numValue value, and finds the EMA with a smoothing factor of 2
after that. isApogee starts returns false before 5 seconds or if there
are fewer than 20 points in the array. Since the array restarts after
writing to the SD card every time, this means that there will be
approximately 600 ms after each write that apogee will not be found.

A design choice that must be made is how to deal with resetting the
array every so often, resulting in there not being previous values to
compare against for isApogee. One solution could be to reset to a number
of points before the end of the array, rather than 0, and calculating
the actual index using an offset. This would slightly lower the capacity
and increase computational cost, but shouldn't be significant.