add series resistor to improve accuracy of ILIM resistor
It used to be a 400 ohm resistor, but those are not in stock in JLC's assembly parts. Now it is two 200 ohm resistors in series. It is important to get the values precise, because the maximum rating of the (inferior, but the only one in stock) TPS2113A is 1.25A, which is also the minimum all-fire current on the E-matches. So we need exactly 1.25A, not above, and not below.